+++ adbenc.h
@@ -9,14 +9,15 @@
  *
  */
 #pragma once
+#include <stdint.h>
 
-void adb_encode(unsigned* inout_state,
+void adb_encode(uint8_t* inout_state,
                 char** inout_enc,
                 char* encend,
                 const char** inout_in,
                 const char* inend);
 
-void adb_decode(unsigned* inout_state,
+void adb_decode(uint8_t* inout_state,
                 char** inout_dec,
                 char* decend,
                 const char** inout_in,
